requirements/requirements.txt
aioredis==1.3.1
appnope==0.1.0
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
backcall==0.1.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
daphne==2.4.1
decorator==4.4.2
dj-database-url==0.5.0
Django==3.0.4
django-cors-headers==3.2.1
django-extensions==2.2.8
django-extra-fields==2.0.3
djangorestframework==3.11.0
drf-yasg==1.17.1
hiredis==1.0.1
html5lib==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
inflection==0.3.1
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.16.0
Jinja2==2.11.1
MarkupSafe==1.1.1
msgpack==0.6.2
packaging==20.3
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.4
psycopg2==2.8.4
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.5.2
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-decouple==3.3
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
Twisted==19.10.0
txaio==20.1.1
uritemplate==3.0.1
urllib3==1.25.8
validate-docbr==1.6.1
wcwidth==0.1.8
webencodings==0.5.1
zope.interface==4.7.1
